config Qt
