+++ src/main.cpp
@@ -1,6 +1,12 @@
 #include <QApplication>
+#include "vue/vueAccueil.h"
 
 int main(int argc, char* argv[]) {
     QApplication app(argc, argv);
+
+    VueAccueil w;
+    w.show();
+
+    return app.exec();
 
 }
